+++ CrossPlatform/TextRenderer.cpp
@@ -348,7 +348,7 @@ int TextRenderer::Render(MultiviewGraphicsDeviceContext& deviceContext, float* x
 	float ht = fontScale * float(GetDefaultTextHeight());
 	//renderPlatform->SetStandardRenderState(deviceContext,crossplatform::STANDARD_ALPHA_BLENDING);
 
-	int n = 0;
+	uint n = 0;
 	crossplatform::StructuredBuffer<FontChar>& f = fontChars[deviceContext.platform_context];
 	FontChar* charList = f.GetBuffer(deviceContext);
 
@@ -416,7 +416,7 @@ int TextRenderer::Render(MultiviewGraphicsDeviceContext& deviceContext, float* x
 			_x += f.pixel_width * fontScale + 1;
 		}
 	}
-	n /= viewCount;
+	n /= static_cast<uint>(viewCount);
 	constantBuffer.numChars = n;
 	if (n > 0)
 	{
